Updated SubRace and Race Property Changes
Updated the SubRace and Race property change so that if the racial bonuses cause the modifier to increase the increases will take place and display on screen.  Additionally modified the selection for SubRace to increase the displayed value of the stat and not just display in the Racial bonus section.